Add files via upload
This repository is under the MIT License, which applies to the code and scripts herein.  
The hash of "Book.rar"(a locked archive of copyrighted work) is provided as a timestamped proof of ownership.  (book_rar_hash.txt)
Note: The content inside Book.rar remains under full copyright and is not subject to the MIT License.